Backend_Flask/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Flask_CQLAlchemy==2.0.0
Flask-APScheduler== 1.13.1 
python-kasa